Got code working, onto worse things
Help
